add bootstrap userpage adminpage \delete = y \add = n \edit = n \show = y

+++ src/main/java/ru/kata/spring/boot_security/demo/servise/RoleService.java
@@ -6,10 +6,10 @@ import ru.kata.spring.boot_security.demo.models.Role;
 import java.util.List;
 
 public interface RoleService {
-    public Role findById(int id);
+    public Role findById(long id);
     public void save(Role role);
-    public void update(int id, Role updRole);
-    public void delete(int id);
+    public void update(long id, Role updRole);
+    public void delete(long id);
 
     public List<Role> findAll();
